pyproject.toml
[project]
name = "articles-searcher-agb"
version = "0.1.0"
description = ""
authors = [
    {name = "a.dolgov",email = "dolgov_am@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "telebot (>=0.0.5,<0.0.6)",
    "pandas (>=2.2.3,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "imaplib2 (>=3.6,<4.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
